some code review changes

+++ pkg/sync/gossip/bit_set.go
@@ -4,19 +4,19 @@ type bitSet struct {
 	array []byte
 }
 
-func (bs *bitSet) set(k int) {
-	bs.array[k>>3] |= (1 << uint32(k&7))
+func (bs *bitSet) set(k uint32) {
+	bs.array[k>>3] |= (1 << (k & 7))
 }
 
-func (bs *bitSet) test(k int) bool {
-	return (bs.array[k>>3] & (1 << uint32(k&7))) != 0
+func (bs *bitSet) test(k uint32) bool {
+	return (bs.array[k>>3] & (1 << (k & 7))) != 0
 }
 
 func (bs *bitSet) toSlice() []byte {
 	return bs.array
 }
 
-func newBitSet(n int) *bitSet {
+func newBitSet(n uint32) *bitSet {
 	return &bitSet{
 		array: make([]byte, (n+7)>>3),
 	}
